pyproject.toml
[project]
name = "pynote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ctkmessagebox>=2.7",
    "customtkinter>=5.2.2",
    "pillow>=11.2.1",
    "pronotepy>=2.14.3",
    "tkchart>=2.1.6",
]
